Clean up API, change tests.
Moved the API entirely to items, as there doesn't seem to be a concept of a user
in the stories.

All the tests pass, although item pagination links simply return the index.
