+++ app/controllers/api/v1/items_controller.rb
@@ -2,17 +2,14 @@
 class Api::V1::ItemsController < ApplicationController
   respond_to :json
 
-  before_filter :validate_request!
+  before_filter :validate_request!, :except => [:index, :show]
 
   def index
-    unless @user = User.find_by_display_name(params[:display_name])
-      render :json => {
-             error: "user #{params[:display_name]} not found"
-             },
-             :status => :not_found
+    if @user = User.find_by_display_name(params[:display_name])
+      @items = Item.find_all_by_poster_id(@user.id)
+    else
+      error(:not_found, "user #{params[:display_name]} does not exist.")
     end
-
-    @items = Item.find_all_by_poster_id(@user.id)
 
     respond_with do |format|
       format.json
@@ -20,14 +17,12 @@ class Api::V1::ItemsController < ApplicationController
   end
 
   def show
-    @item = Item.find(params[:id])
-    @user = User.find_by_display_name(params[:display_name])
+    unless @user = User.find_by_display_name(params[:display_name])
+      error(:not_found, "user #{params[:display_name]} does not exist.")
+    end
 
-    unless @user
-      render :json => {
-             error: "user #{params[:display_name]} not found"
-             },
-             :status => :not_found
+    unless @item = Item.find(params[:id])
+      error(:not_found, "item #{params[:id]} does not exist.")
     end
 
     respond_with do |format|
@@ -40,7 +35,7 @@ class Api::V1::ItemsController < ApplicationController
       @post = build_post_from(params, post_type)
 
       if @post.save
-        success(:created, "#{post_type} created successfully")
+        success(:created, "#{post_type} created successfully", @post)
       else
         error(:unprocessable_entity, "#{post_type} could not be created")
       end
@@ -69,7 +64,9 @@ class Api::V1::ItemsController < ApplicationController
                    :location => '')
   end
 
-  def success(type, msg = nil)
+  def success(type, msg = nil, post = nil)
+    @item = post
+
     respond_with({ :success  => "#{ msg || type }" }.to_json,
                    :status   => type,
                    :location => '')
